fix(backend): include authenticated user's rating in RecipeDetailsDto
