+++ backend/src/main/java/com/group1/cuisines/config/JwtAuthenticationFilter.java
@@ -2,6 +2,8 @@ package com.group1.cuisines.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.group1.cuisines.dao.response.ErrorResponse;
+import com.group1.cuisines.entities.User;
+import com.group1.cuisines.repositories.UserRepository;
 import com.group1.cuisines.services.JwtService;
 import com.group1.cuisines.services.UserService;
 import io.jsonwebtoken.ExpiredJwtException;
@@ -17,17 +19,24 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 
 @Component
 @RequiredArgsConstructor
 public class JwtAuthenticationFilter extends OncePerRequestFilter {
 
+    private final UserRepository userRepository;
     private final JwtService jwtService;
-    private final UserService userService;
     private final ObjectMapper objectMapper;
+
+    private static final Logger logger = LoggerFactory.getLogger(UserService.class);
 
     @Override
     protected void doFilterInternal(
@@ -57,9 +66,7 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
                     StringUtils.isNotEmpty(userEmail) &&
                             SecurityContextHolder.getContext().getAuthentication() == null
             ) {
-                UserDetails userDetails = userService
-                        .userDetailsService()
-                        .loadUserByUsername(userEmail);
+                UserDetails userDetails = userDetailsService().loadUserByUsername(userEmail);
                 if (jwtService.isTokenValid(jwt, userDetails)) {
                     SecurityContext context =
                             SecurityContextHolder.createEmptyContext();
@@ -87,5 +94,21 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 
         }
         filterChain.doFilter(request, response);
+    }
+
+    public UserDetailsService userDetailsService() {
+        return new UserDetailsService() {
+
+            @Override
+            public UserDetails loadUserByUsername(String usernameOrEmail) {
+                logger.debug("Attempting to find user by email or username: {}", usernameOrEmail);
+
+                User user = userRepository.findByEmailOrUsername(usernameOrEmail, usernameOrEmail)
+                        .orElseThrow(() -> new UsernameNotFoundException("User not found with username or email : " + usernameOrEmail));
+
+                return user;
+            }
+
+        };
     }
 }
